Website Images and Link
Link - https://www.dropbox.com/s/4dky35v517y4ntl/Web%20Development%20Website%20Assignment.zip?dl=0

The PDF uploaded is a montage of each page on a website I created in the first year of University for one of my modules. As it is not published, you need to download the .zip from the dropbox link posted above. This will allow you to view the website offline on your machine.